plusTipMergeSubtracks returns two data matrices - one where growth fgaps are given a new identifier (5), and one where they are merged with the flanking growth phases (so matrix only has growth, fgaps, and bgaps, similar to matrix before some fgaps were reclassified as growth).
plusTipPostTracking now stores the former matrix but uses the latter matrix to calculate statistics. Histogram function is now a subfunction instead of a standalone, and the whole meta folder is cleared first if post-processing is being re-run.

plusTipGetSubtrackCoords now allows you to specify whether or not to use the merged or reclassified matrix (result of plusTipMergeSubtracks), which matters if indices are given. 

plusTipPlotTracks now shows growth fgaps as solid green line.

plusTipSpeedMovie now uses circles, triangles, and squares to represent growths, fgaps, and bgaps respectively.

plusTipParamPlot now has labels of fgap/bgap instead of pause/shrinkage

git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/mtTracker/main@3595 0bcb941b-74d1-0310-bb26-e6636707f451