Deepak -- intravital -- stable -- DNADamageAnalysis code
Changes to old code:

--------------------------------

-- Added a GUI feature to CellSegmentationQualityAnnotator.m, CellPatternAnnotator.m, CellStateAnalyzer.m

Now u can click on a cell with the mouse and that cell will become the active cell and also the three ortho slice image panels (XY, XZ, YZ) will be updated accordingly.

-- ImarisDataVisualizer.m 

Fixed the spot visualization part

--------------------------------

New additions:

-- DNADamageAnalyzer.m, segmentFociInsideNuclei.m

Developed a tool for detecting foci/puncta in the channel tagging 53BP1 DNA repair protein in HT-1080 cells. The tool segments the nuclei and then find and detects foci inside the nuclei. The goal is to study the potency of platinum compound drugs (ex: cisplatin) at inducing 53BP1 puncta.

-- batchDNADamageAnalysis.m, performDNADamageAnalysis.m

These are scripts for running DNADamageAnalysis on the cluster (orchestra).


git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/intravital/main@11377 0bcb941b-74d1-0310-bb26-e6636707f451