Deepak -- intravital -- stable -- DNADamageAnalysis
Now i store all the channels in the outputted mat file and type cast them to uint16 so the file size is small.

In segmentFociInsideNuclei.m, i replaced median smoothing [1, 1, 0] with gaussian smoothing with sigma = half pixel size in XY. This is a bit more sensitive to the detection of tiny puncta.





git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/intravital/main@11396 0bcb941b-74d1-0310-bb26-e6636707f451