Deepak --- intravital --- work in progress
>> ComputeDNADamageColocalizationMeasures is now more generic.
It takes the 53bp1 image and an image to compute its colocalization with
which could me the drug or macrohage markers.

>> performDNADamageAnalysis

The drug and macrophage channels may or may not be present now but the
53BP1 channel is mandatory

Now the colocalization with the drug and macrophage marker channels are
computed separately and outputted in separate files depending on whether
these markers are present in the input data or not.

batchDNADamageAnalysis was changed accordingly to work with the changes in
the interface of performDNADamageAnalysis

>> segmentCellsInIntravitalData

Testing this on other in vivo data. So i am debugging parts of it.