added flag to put a negative exit flag (-20) for those filopodia I know projected forward in an incorrect direction- as they are relatively rare just currently filter out as low confidence for now- but eventually troubleshoot.
added a test for if the u = the mindistfilo
if so likely enough points were used in the fitting- increase the fit to fit to the full xyCoords sampled. So far this fixes fits quite well.
examples that benefited were filo 18,21,28,31- so most of these benefited quite a bit from this