Fix To Find Best Endpoint When there are Multiple Potential Entrance Points from a Spanning Ridge Candidate.
Small problem in function in that in the interpolation to the edge it would interpolate the shortest distance from the boundary – if there were two competing paths it did not test the path but simply choose the first one which is obviously incorrect.

The fix was to test to simply find the shortest distance from the original cluster first then interpolate to the edge if necessary after the best EP was found. - Note Should come back to optimize function further before release -but for now main problem solved.

Fix at line 399
See Task Fix Switch Entry Point 20141205 Data Cell02 Frame 111 for details